requirements.txt
watchdog
pdfplumber
Flask
dotenv
mysql-connector-python
bcrypt